parcelviz/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parcelviz"
version = "0.1.0"
description = "Parcel visualization pipeline with REST API and CLI."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "ParcelScope", email = "dev@example.com" },
]
license = { text = "Proprietary" }
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.5",
  "pydantic-settings>=2.0",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "requests>=2.31",
  "requests-cache>=1.2",
  "matplotlib>=3.8",
  "Pillow>=10.0",
  "contextily>=1.5",
  "mercantile>=1.2",
  "rasterio>=1.3",
  "scipy>=1.11",
  "numpy>=1.26",
  "python-dateutil>=2.8",
  "typer[all]>=0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "responses>=0.25",
  "respx>=0.20",
  "httpx>=0.26",
  "mypy>=1.8",
  "ruff>=0.2",
]

[project.scripts]
parcelviz = "parcelviz.cli:app"

[tool.setuptools]
package-dir = {"" = "parcelviz"}

[tool.setuptools.packages.find]
where = ["parcelviz"]
